feat: cut out the middle main
