+++ packages/ebsr/configure/src/index.js
@@ -1,32 +1,83 @@
-import Main from './main';
+import { ModelUpdatedEvent } from '@pie-framework/pie-configure-events';
+import MultipleChoiceConfigure from '@pie-element/multiple-choice/configure/lib';
+
+import debug from 'debug';
+
+const log = debug('pie-elements:ebsr:configure');
 
 export default class EbsrConfigure extends HTMLElement {
-  static getMain = ()  => document.querySelector('ebsr-main');
-
-  static defineMain() {
-    if (!customElements.get('ebsr-main')) {
-      customElements.define('ebsr-main', Main);
+  static defineMultipleChoice() {
+    if(!customElements.get('multiple-choice-configure')){
+      customElements.define('multiple-choice-configure', MultipleChoiceConfigure);
     }
   }
 
-  static setModel(m) {
-    const main = EbsrConfigure.getMain();
-    main.model = m;
-  }
+  static getParts = () => ({
+    partA: document.getElementById('part-a-configure'),
+    partB: document.getElementById('part-b-configure'),
+  });
 
   constructor() {
     super();
 
-    EbsrConfigure.defineMain();
+    EbsrConfigure.defineMultipleChoice();
   }
 
   set model(m) {
-    this._render();
+    this._model = m;
 
-    EbsrConfigure.setModel(m);
+    this._render();
+    this.setupModel();
+  }
+
+  setupModel() {
+    const { partA, partB } = EbsrConfigure.getParts();
+
+    if (partA && partB) {
+      this.setupPart(partA, 'partA');
+      this.setupPart(partB, 'partB');
+    }
+  }
+
+  setupPart(part, key) {
+    this.initiatePart(part, key);
+    this.captureModelUpdated(part, key);
+  }
+
+  initiatePart(part, key) {
+    part.model = this._model[key];
+
+    this.dispatchModelChanged();
+  }
+
+  dispatchModelChanged() {
+    log('[onModelChanged]: ', this._model);
+
+    this.dispatchEvent(new ModelUpdatedEvent(this._model, false));
+  }
+
+  captureModelUpdated(part, key) {
+    part.addEventListener('model.updated', event => {
+      event.stopImmediatePropagation();
+      this.repackModel(event.update, key);
+    });
+  }
+
+  repackModel(partModel, key) {
+    this._model =  {
+      ...this._model,
+      [key]: partModel
+    };
+
+    this.dispatchModelChanged();
   }
 
   _render() {
-    this.innerHTML = `<ebsr-main></ebsr-main>`;
+    this.innerHTML = `
+      <div>
+        <multiple-choice-configure id="part-a-configure"></multiple-choice-configure>
+        <multiple-choice-configure id="part-b-configure"></multiple-choice-configure>
+      </div>
+    `;
   }
 }
